requirements.txt
aiogram
python-dotenv
psycopg[binary]